requirements.txt
albumentations==1.1.0
Flask==2.0.2
ipython==8.0.0
matplotlib==3.4.3
numpy==1.20.2
opencv-python==4.5.4.60
opencv_contrib_python==4.5.4.60
opencv-python-headless==4.5.4.60
pandas==1.3.4
Pillow==9.0.0
protobuf==3.19.3
PyYAML==6.0
requests==2.24.0
scipy==1.7.2
seaborn==0.11.2
setuptools==52.0.0.post20210125
thop==0.0.31.post2005241907
torch==1.8.0
torchvision==0.9.1
tqdm==4.51.0
